Javadoc and small chnage to discard

+++ galasa-managers-parent/galasa-managers-cloud-parent/dev.galasa.docker.manager/src/main/java/dev/galasa/docker/internal/DockerVolumeImpl.java
@@ -7,8 +7,12 @@ import org.apache.commons.logging.LogFactory;
 
 import dev.galasa.docker.DockerManagerException;
 import dev.galasa.docker.IDockerVolume;
-import dev.galasa.framework.spi.IFramework;
 
+/**
+ * A implementation of the docker volumes that will be brought up on the engine
+ * 
+ * @author James Davies
+ */
 public class DockerVolumeImpl implements IDockerVolume {
     private final String volumeName;
     private final String mountPath;
@@ -17,6 +21,15 @@ public class DockerVolumeImpl implements IDockerVolume {
 
     private static final Log logger = LogFactory.getLog(DockerVolumeImpl.class);
 
+    /**
+     * Constructor that determines the nature of the volume (readOnly or not), and provisions or ensures
+     * the volumes exsists.
+     * 
+     * @param volumeName
+     * @param mountPath
+     * @param engine
+     * @throws DockerManagerException
+     */
     public DockerVolumeImpl(String volumeName, String mountPath, DockerEngineImpl engine)
             throws DockerManagerException {
 
@@ -44,21 +57,41 @@ public class DockerVolumeImpl implements IDockerVolume {
         logger.info("Existing volume found.");
     }
 
+    /**
+     * Returns volume name
+     * 
+     * @return volumeName
+     */
     @Override
-    public String getVoumeName() {
+    public String getVolumeName() {
         return this.volumeName;
     }
 
+     /**
+     * Returns mount path
+     * 
+     * @return mountPath
+     */
     @Override
     public String getMountPath() {
         return this.mountPath;
     }
 
+     /**
+     * Returns readonly state
+     * 
+     * @return readOnly
+     */
     @Override
     public boolean readOnly() {
         return this.readOnly;
     }
 
+     /**
+     * Checks the docker engine for a specific named volume
+     * 
+     * @return boolean exists
+     */
     public boolean doesVolumeExist() throws DockerManagerException {
         if (engine.getVolume(this.volumeName) != null) {
             return true;   
@@ -66,8 +99,17 @@ public class DockerVolumeImpl implements IDockerVolume {
         return false;
     }
 
+    /**
+     * Discard a volume. This should only be called for galasa created volumes.
+     * 
+     * @throws DockerManagerException
+     */
     public void discard() throws DockerManagerException {
-        engine.deleteVolume(this.volumeName);
+        if (readOnly) {
+            logger.error("Not deleted, not a Galasa volume!");
+        } else {
+            engine.deleteVolume(this.volumeName);
+        }
     }
     
 }
